+++ vpin-connector-iscored/src/main/java/de/mephisto/vpin/connectors/iscored/IScored.java
@@ -104,7 +104,7 @@ public class IScored {
       }
     }
     catch (Exception e) {
-      LOG.error("Failed to load iScored Game Room: {}", e.getMessage(), e);
+      LOG.error("Failed to load iScored Game Room: {}", e.getMessage());
     }
 
     return null;