+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/recorder/EmulatorRecorderJob.java
@@ -115,6 +115,7 @@ public class EmulatorRecorderJob extends FrontendRecorderJob {
       finally {
         //this will kill the emulators too
         frontend.killFrontend();
+        gameRecorder.finalizeRecordings();
       }
     }
     LOG.info("Recordings for " + recordingDataSummary.size() + " games finished.");