+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/recorder/FrontendRecorderJob.java
@@ -110,6 +110,7 @@ public class FrontendRecorderJob implements Job {
       }
       finally {
         frontend.endFrontendRecording();
+        gameRecorder.finalizeRecordings();
       }
     }
     LOG.info("Recordings for " + recordingDataSummary.size() + " games finished.");